Fix channel wraparound
When going beyound the cahhnels, divibly was supposed to wraparound to
the other end, unfortunately this wasn't working. This was due to
passing in chan_idx to play_channel() when in fact chan_idx is a global
variable, it seems this is a bad idea, causing chan_idx to seemingly
have two values the correct one and the out of bounds one.

The fix is simply to remove the argument from play_channel()

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>