Add the ability to search for channels
A Gtk Entry/Completion widget that can be used to search through the
list of channels has been added.

A vertical box was added to the main window and moved the drawing area
into there and then a Gtk Entry added to the bottom part.

The text entry is hidden by default and can be activated with the '/'
key. It can be dismissed with Esc.

At startup a Gtk List Store is created that holds the channel name and
index into the channels array for that channel.

When a channel is selected we get the index value and use that to set
chan_idx then initiate play_channel(), we then clear and re-hide the
text entry.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>